requirements.txt
argparse
colorama
dashscope
opencv-python
pyyaml
requests
google-generativeai
Pillow 